remove plotting logic (which should stay in analysis) from calculations (which should be either in equ or neq)
